18.c

+++ 18.c
@@ -2,7 +2,7 @@
 int main()
 {
     int n;
-    unsigned long long int fact = 1;
+    int fact = 1;
     printf("Enter the value of n:");
 
     scanf("%d", &n);
@@ -11,6 +11,6 @@ for(int i=1;i<=n;i++)
 {
 fact*=i;
 }
-printf("factorial of a number is %llu",fact);
+printf("factorial of a number is %d",fact);
     return 0;
 }
